LPD-86157 Move into util class instead, and update usages
