--- modules/dxp/apps/ai-hub/ai-hub-api/src/main/java/com/liferay/ai/hub/agent/OAuth2ApplicationIdResolver.java
@@ -1,20 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.ai.hub.agent;
-
-import com.liferay.portal.kernel.security.auth.PrincipalException;
-
-import jakarta.servlet.http.HttpServletRequest;
-
-/**
- * @author Christopher Kian
- */
-public interface OAuth2ApplicationIdResolver {
-
-	public long resolve(HttpServletRequest httpServletRequest)
-		throws PrincipalException;
-
-}
